requirements.txt
ijson
pokebasex
